+++ src/AzzyBot-Next/Commands/DebugCommands.cs
@@ -2,12 +2,14 @@
 using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
+using AzzyBot.Commands.Choices;
 using AzzyBot.Logging;
 using AzzyBot.Services;
 using AzzyBot.Utilities.Encryption;
 using DSharpPlus.Commands;
 using DSharpPlus.Commands.ArgumentModifiers;
 using DSharpPlus.Commands.ContextChecks;
+using DSharpPlus.Commands.Processors.SlashCommands.ArgumentModifiers;
 using DSharpPlus.Entities;
 using Microsoft.Extensions.Logging;
 
@@ -41,21 +43,21 @@ public sealed class DebugCommands
         public async ValueTask DebugTriggerExceptionAsync
             (
             CommandContext context,
-            [Description("Enable to defer the message before throwing the exception.")] bool throwAfterDefering = false,
-            [Description("Enable to throw the exception after a reply was already made.")] bool afterReply = false
+            [Description("Enable to defer the message before throwing the exception."), SlashChoiceProvider<BooleanEnableDisableStateProvider>] int throwAfterDefering = 0,
+            [Description("Enable to throw the exception after a reply was already made."), SlashChoiceProvider<BooleanEnableDisableStateProvider>] int afterReply = 0
             )
         {
             ArgumentNullException.ThrowIfNull(context, nameof(context));
 
             _logger.CommandRequested(nameof(DebugTriggerExceptionAsync), context.User.GlobalName);
 
-            if (throwAfterDefering)
+            if (throwAfterDefering is 1)
                 await context.DeferResponseAsync();
 
-            if (afterReply && !throwAfterDefering)
+            if (afterReply is 1 && throwAfterDefering is not 1)
                 await context.RespondAsync("This is a debug reply");
 
-            if (afterReply && throwAfterDefering)
+            if (afterReply is 1 && throwAfterDefering is 1)
                 await context.EditResponseAsync("This is a debug reply edit");
 
             throw new InvalidOperationException("This is a debug exception");