+++ src/AzzyBot-Next/Utilities/Records/AzuraCast/AzuraSystemLogsRecord copy.cs
@@ -1,0 +1,9 @@
+using System.Text.Json.Serialization;
+
+namespace AzzyBot.Utilities.Records.AzuraCast;
+
+public sealed record AzuraSystemLogRecord
+{
+    [JsonPropertyName("contents")]
+    public required string Content { get; init; }
+}